+++ tools/libexec/daemons/mqtt-monitor-daemon.pl
@@ -273,18 +273,20 @@ sub doMatched {
 				if( $enabled ){
 					$enabledCount += 1;
 					my $jsonable = TTP::JSONable->new( $ep, $do );
-					my $command = TTP::commandByOS([], { jsonable => $jsonable, withCommand => true });
-					if( $command ){
-						my $res = TTP::commandExec( $command, {
-							macros => {
-								TOPIC => $topic,
-								PAYLOAD => $payload
+					my $commands = TTP::commandByOS([], { jsonable => $jsonable });
+					if( $commands && scalar( @{$commands} )){
+						foreach my $cmd ( @{$commands} ){
+							my $res = TTP::commandExec( $cmd, {
+								macros => {
+									TOPIC => $topic,
+									PAYLOAD => $payload
+								}
+							});
+							if( $res->{success} ){
+								$successCount += 1;
+							} else {
+								$failedCount += 1;
 							}
-						});
-						if( $res->{success} ){
-							$successCount += 1;
-						} else {
-							$failedCount += 1;
 						}
 					} else {
 						msgVerbose( "$topic: action n° $totalCount doesn't have a command" );